move 'renderpage' to generic blog post
