+++ src/advancedKubectl/kubectl-productivity-related.md
@@ -1,0 +1,6 @@
+# That's all folks!
+
+If you enjoyed this article, you might find the following articles interesting:
+
+- [How to practice chaos engineering and reduce costs](/blog/kubernetes-spot-instances) with Kubernetes and spot instances.
+- [Scaling Microservices with Message Queues, Spring Boot and Kubernetes.](/blog/scaling-spring-boot-microservices) Learn how to use the Horizontal Pod Autoscaler to resize your fleet of applications dynamically.
