+++ src/chaosEngineering/chaos-engineering-related.md
@@ -1,0 +1,6 @@
+# That's all folks!
+
+If you enjoyed this article, you might find the following articles interesting:
+
+- [3 simple tricks for smaller Docker images](/blog/smaller-docker-images) and learn how to build and deploy Docker images quicker.
+- [Scaling Microservices with Message Queues, Spring Boot and Kubernetes.](/blog/scaling-spring-boot-microservices) Learn how to use the Horizontal Pod Autoscaler to resize your fleet of applications dynamically.
