+++ src/genericBlogPost.tsx
@@ -1,0 +1,115 @@
+import * as React from 'react'
+import { State, getPages, getOpenGraph, getBlogPosts, getAuthors, getBlogPostMarkdownBlocks } from './store'
+import { Page } from './store/websiteReducer'
+import { Subscribe } from './layout.v2'
+import { Head, Html, Body, Navbar, OpenGraph, Footer } from './layout.v3'
+import { format } from 'date-fns'
+import { join } from 'path'
+import { read } from './files'
+import { JsonLd } from 'react-schemaorg'
+import { BlogPosting } from 'schema-dts'
+import { renderToJsx, toMdast } from './markdown'
+import { Author } from './article.v2'
+import { transform } from './markdown/utils'
+import { selectAll } from 'unist-util-select'
+import * as Mdast from 'mdast'
+import { mdast2Jsx } from './markdown/jsx'
+
+export async function renderPage(pageMeta: Page, state: State) {
+  const page = getPages(state).find(it => it.id === pageMeta.id)!
+  const openGraph = getOpenGraph(state).find(it => it.pageId === pageMeta.id)
+  if (!openGraph) {
+    throw new Error('The page does not have an open graph.')
+  }
+  const blog = getBlogPosts(state).find(it => it.pageId === pageMeta.id)
+  if (!blog) {
+    throw new Error('The page is not a blog post page.')
+  }
+  const author = getAuthors(state).find(it => it.id === blog.authorId)
+  if (!author) {
+    throw new Error('The blog post does not have an author attached')
+  }
+  const currentAbsoluteUrl = `${state.config.protocol}://${join(state.config.hostname, page.url)}`
+  const extraBlocks = getBlogPostMarkdownBlocks(state).filter(it => it.blogPostId === blog.id)
+  const [content, ...blocks] = await Promise.all([
+    read(blog.content!),
+    ...extraBlocks.map(it => it.content).map(it => read(it)),
+  ])
+  return (
+    <Html>
+      <Head title={page.title} description={page.description}>
+        {openGraph ? (
+          <OpenGraph
+            title={openGraph.title}
+            description={openGraph.description}
+            image={openGraph.image}
+            currentAbsoluteUrl={currentAbsoluteUrl}
+          />
+        ) : null}
+        <link rel='stylesheet' href='node_modules/tachyons/css/tachyons.css' />
+        <link rel='stylesheet' href='assets/style.css' />
+        <JsonLd<BlogPosting>
+          item={{
+            '@context': 'https://schema.org',
+            '@type': 'BlogPosting',
+            headline: blog.title,
+            image: `${openGraph.image.props.src}`,
+            author: {
+              '@type': 'Person',
+              name: author.fullName,
+            },
+            publisher: {
+              '@type': 'Organization',
+              name: 'Learnk8s',
+              logo: {
+                '@type': 'ImageObject',
+                url: `assets/learnk8s_logo_square.png`,
+              },
+            },
+            url: currentAbsoluteUrl,
+            datePublished: blog.publishedDate,
+            dateModified: blog.lastModifiedDate || blog.publishedDate,
+            description: blog.description,
+            mainEntityOfPage: {
+              '@type': 'SoftwareSourceCode',
+            },
+          }}
+        />
+      </Head>
+      <Body>
+        <div className='white mb4 mb5-ns'>
+          <Navbar />
+        </div>
+        <div className='tc mb4 db mw4 center'>
+          <Author name={author.fullName} avatar={author.avatar} link={author.link} />
+        </div>
+        <article className='ph3 pt0 pb4 mw7 center'>
+          <h1 className='navy tc f2 f1-ns'>{blog.title}</h1>
+          <p className='f7 black-60 tc ttu'>Published in {format(new Date(blog.publishedDate), 'MMMM yyyy')}</p>
+          {blog.lastModifiedDate ? (
+            <p className='f7 black-60 tc ttu b'>
+              <img src='assets/tick.svg' alt='Tick' className='w1 h1 v-mid' /> Updated in{' '}
+              {format(new Date(blog.lastModifiedDate), 'MMMM yyyy')}
+            </p>
+          ) : null}
+          <img src={openGraph.image.props.src} className='pt2' alt={openGraph.image.props.alt} />
+          <hr className='w3 center b--navy mv4 mb5-ns' />
+          {renderToJsx(content)}
+          {blocks.map(it => {
+            const mdast = toMdast(it)
+            bumpHeadings(mdast.children, 1)
+            return transform({ type: 'root' as const, children: mdast.children }, mdast2Jsx())
+          })}
+          <Subscribe identifier={blog.id} />
+        </article>
+        <Footer />
+      </Body>
+    </Html>
+  )
+}
+
+function bumpHeadings(children: Mdast.Content[], amount: number): void {
+  selectAll<Mdast.Heading>('heading', { type: 'root', children }).forEach(heading => {
+    heading.depth = heading.depth + amount
+  })
+}
